# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server Petstore server.",
        "title": "Swagger Example API12222",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "petstore.swagger.io",
    "basePath": "/v1",
    "paths": {
        "/api/users": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "获取所有用户",
                "responses": {
                    "200": {
                        "description": "{\"RetCode\":0,\"UserInfo\":{},\"Action\":\"GetAllUserResponse\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/users/{name}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "获取单个用户",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"RetCode\":0,\"UserInfo\":{},\"Action\":\"GetOneUserResponse\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    }
}